Mise en place Contraintes SQL Unique pour : MenuItem.name et Facture.commande

+++ src/main/java/com/example/tp_resto/service/FactureServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -19,50 +20,90 @@ public class FactureServiceImpl implements FactureService {
         this.factureRepo = factureRepo;
     }
 
+    //Create
     @Override
     public Facture saveFacture(Facture facture) {
+        // Check if a Facture with the same commandeId already exists
+        if (factureRepo.existsByCommandeId(facture.getCommande().getId())) {
+            throw new IllegalArgumentException("Facture with commande ID " + facture.getCommande().getId() + " already exists.");
+        }
         return factureRepo.save(facture);
     }
 
+    //Read
+    @Override
+    public Facture findById(int id) {
+        return factureRepo.findById(id)
+                .orElseThrow(() -> new FactureNotFoundException("Facture No " + id + " Not Found"));
+    }
+
+    @Override
+    public List<Facture> findAll() {
+
+        List<Facture> maListe = factureRepo.findAll();
+        if(maListe.isEmpty()) {
+            throw new FactureNotFoundException("Aucune factures trouvées");
+        } else  return maListe;
+    }
+
+
+    @Override
+    public Facture findByDate(Date date) { //TODO
+        return null;
+    }
+
+    //Update
     @Override
     public Facture updateFacture(int id, Facture facture) {
 
         Optional<Facture> factureOptional = factureRepo.findById(id);
-        Facture existingFacture = null;
+
         if (factureOptional.isPresent()) {
-            existingFacture = factureOptional.get();
+            Facture existingFacture = factureOptional.get();
 
-            existingFacture.setCommande(facture.getCommande());
-            existingFacture.setStatus(facture.getStatus());
-            existingFacture.setBillTime(facture.getBillTime());
-            existingFacture.setPrix(facture.getPrix());
+            //On verifie si on essaie de changer l'assignation a une commandes différente
+            boolean facturesAssignesADesCommandesDifferentes = !existingFacture.getCommande().getId()
+                    .equals(facture.getCommande().getId());
 
-            return factureRepo.save(existingFacture);
-        } else {
-            throw new FactureNotFoundException("Facture Not Found");
+
+            boolean factureEstDejaAssigneAUneCommande = factureRepo.existsByCommandeId(facture.getCommande().getId());
+
+            if (factureEstDejaAssigneAUneCommande && facturesAssignesADesCommandesDifferentes) {
+                throw new IllegalArgumentException("Facture reliée avec commande ID " + facture.getCommande().getId()
+                        + " existe déja.");
+            }
+
+                existingFacture.setCommande(facture.getCommande());
+                existingFacture.setStatus(facture.getStatus());
+                existingFacture.setBillTime(facture.getBillTime());
+
+                return factureRepo.save(existingFacture);
+            } else {
+                throw new FactureNotFoundException("Facture No " + id + " Not Found");
+            }
         }
 
-    }
 
-    @Override
-    public Facture findById(int id) {
-        return factureRepo.findById(id)
-                .orElseThrow(() -> new FactureNotFoundException("Facture Not Found"));
-    }
-
-
-    @Override
-    public Facture findByDate(Date date) {
-        return null;
-    }
-
+    //Delete
     @Override
     public void deleteFacture(Facture facture) {
+        Optional<Facture> factureOpt = factureRepo.findById(facture.getId());
+        if(factureOpt.isPresent()){
+
         factureRepo.delete(facture);
+        } else {
+            throw new FactureNotFoundException("Facture No " + facture.getId() + " Not Found");
+        }
     }
 
     @Override
     public void deleteFactureById(int id) {
-        factureRepo.deleteById(id);
+        Optional<Facture> factureOpt = factureRepo.findById(id);
+        if(factureOpt.isPresent()){
+
+            factureRepo.deleteById(id);
+        } else {
+            throw new FactureNotFoundException("Facture No " + id + " Not Found");
+        }
     }
 }
